requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
pywinpty>=2.0.0;sys_platform=="win32"
